django/resume_uploader/requirements.txt
django
djangorestframework
pillow
django-cors-headers
psycopg2-binary
python-dotenv